+++ src/js-joda-extra.js
@@ -2,6 +2,14 @@
  * @copyright (c) 2016, Philipp Thuerwaechter & Pattrick Hueper
  * @license BSD-3-Clause (see LICENSE.md in the root directory of this source tree)
  */
-import extraDefault from './main';
 
-export default extraDefault;
+import { use } from 'js-joda';
+
+import { Interval } from './Interval';
+import plug from './plug';
+
+use(plug);
+
+export {
+    Interval,
+};