+++ test/mainTest.js
@@ -4,28 +4,14 @@
  */
 
 import { expect } from 'chai';
-import { use } from 'js-joda';
 
-// since for karma/webpack tests we cannot require the actual webpack entry point,
-// we test using `main` ... the entry point js-joda-extra.js only re-exports the default from
-// main anyway...
-import extra from '../src/main';
-
-let jsJoda;
-
-const testPlugin = (plugin) => {
-    before('use', () => {
-        jsJoda = use(plugin);
-    });
-
-    it('should add Interval to joda', () => {
-        expect(jsJoda.Interval).to.exist;
-    });
-};
+import '../src/js-joda-extra';
+import { Interval } from 'js-joda';
 
 describe('main test', () => {
-    testPlugin(extra);
+    it.skip('should add Interval to joda', () => {
+        // we have to import Interval from js-joda-extra from now on
+        expect(Interval).to.exist;
+    });
 });
 
-// export the testPlugin function to be re-used in pluginTest_mochaOnly
-export default testPlugin;